requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dateutil
werkzeug
pdfplumber
openpyxl
gunicorn
pandas